requirements.txt
requests
paho-mqtt
python-dotenv